+++ tests/testthat/helper.R
@@ -1,3 +1,5 @@
+options("rgdal_show_exportToProj4_warnings"="none")
+
 library("sf")
 library("dplyr")